+++ changelog.md
@@ -43,7 +43,7 @@ For example:
 * Made it possible to extend the app layout.
 * Allow to import more than 15Mb. [949](ttps://gitlab.com/bramw/baserow/-/issues/949)
 * Add the ability to disable the model cache with the new BASEROW_DISABLE_MODEL_CACHE env variable.
-* Applications can be duplicated. [#960](https://gitlab.com/bramw/baserow/-/issues/960)
+* Applications can now be duplicated. [#960](https://gitlab.com/bramw/baserow/-/issues/960)
 
 ### Bug Fixes